Split delegation and default chat agent
Also remove the special case with the delegation agent label and truly support multiple chat agents chiming in.

Fixes https://github.com/eclipsesource/osweek-2024/issues/125
