+++ packages/ai-chat-ui/src/browser/chat-tree-view/chat-view-tree-widget.tsx
@@ -174,6 +174,9 @@ export class ChatViewTreeWidget extends TreeWidget {
         if (!(isRequestNode(node) || isResponseNode(node))) {
             return super.renderNode(node, props);
         }
+        if (isRequestNode(node) && node.request.request.isHidden) {
+            return undefined;
+        }
         return <React.Fragment key={node.id}>
             <div className='theia-ChatNode'>
                 {this.renderAgent(node)}
@@ -194,20 +197,10 @@ export class ChatViewTreeWidget extends TreeWidget {
     private getAgentLabel(node: RequestNode | ResponseNode): string {
         if (isRequestNode(node)) {
             // TODO find user name
-            return 'Me';
+            return 'You';
         }
-
         const agent = node.response.agentId ? this.chatAgentService.getAgent(node.response.agentId) : undefined;
-        const initialAgentlabel = agent?.name ?? 'AI';
-        const labelParts = [initialAgentlabel];
-
-        for (const delegateAgentId of node.response.delegateAgentIds) {
-            const delegateAgent = this.chatAgentService.getAgent(delegateAgentId);
-            const delegateAgentlabel = delegateAgent?.name ?? 'AI';
-            labelParts.push(delegateAgentlabel);
-        }
-
-        return labelParts.join(' > ');
+        return agent?.name ?? 'AI';
     }
     private getAgentIconClassName(node: RequestNode | ResponseNode): string | undefined {
         if (isRequestNode(node)) {
